requirements.txt
discord.py==1.7.3
requests==2.28.1
pymongo==4.2.0
pymongo[tls,srv]==4.2.0